import prod
Former-commit-id: 5e1adb5de4638960ecb473e6aa023ce0c3529a8e

+++ htdocs/bimpcore/scripts/import_products.php
@@ -450,6 +450,7 @@ class importCatalogueFourn{
     public  $dir = '';
     public $sep = ";";
     public $updateSql = true;
+    public $modeProdFourn = "insert";
     
     function majStock($ref, $qty){
         global $db;
@@ -710,7 +711,7 @@ class importCatalogueFourn{
     function truncTableProdFourn()
     {
         global $db;
-        if($this->updateSql)
+        if($this->updateSql && $this->modeProdFourn == "insert")
             $db->query("DELETE FROM " . MAIN_DB_PREFIX . "bimp_product_import_fourn WHERE id_fourn = ".$this->idFourn);
     }
     
@@ -733,9 +734,14 @@ class importCatalogueFourn{
         $marque = addslashes($marque);
         $lib = addslashes($lib);
         $refFabriquant = addslashes($refFabriquant);
-        if($this->updateSql)
-            $db->query("INSERT INTO `" . MAIN_DB_PREFIX . "bimp_product_import_fourn`(id_fourn, `refLdLC`, `codeLdlc`, `pu_ht`, `tva_tx`, `pa_ht`, `marque`, `libelle`, `refFabriquant`, `data`) "
-                . "VALUES (".$this->idFourn.", '" . $refLdlc . "','" . $codeLdlc . "','" . $pu_ht . "','" . $tva_tx . "','" . $pa_ht . "','" . $marque . "','" . $lib . "','" . $refFabriquant . "','" . $data . "')");
+        if($this->updateSql){
+            if($this->modeProdFourn == "insert")
+                $db->query("INSERT INTO `" . MAIN_DB_PREFIX . "bimp_product_import_fourn`(id_fourn, `refLdLC`, `codeLdlc`, `pu_ht`, `tva_tx`, `pa_ht`, `marque`, `libelle`, `refFabriquant`, `data`) "
+                    . "VALUES (".$this->idFourn.", '" . $refLdlc . "','" . $codeLdlc . "','" . $pu_ht . "','" . $tva_tx . "','" . $pa_ht . "','" . $marque . "','" . $lib . "','" . $refFabriquant . "','" . $data . "')");
+            else
+                $db->query("UPDATE `" . MAIN_DB_PREFIX . "bimp_product_import_fourn` SET codeLdlc = '" . $codeLdlc . "', pu_ht = '" . $pu_ht . "', tva_tx = '" . $tva_tx . "', pa_ht = '" . $pa_ht . "',marque = '" . $marque . "', libelle= '" . $lib . "',refFabriquant = '" . $refFabriquant . "',data = '" . $data . "'"
+                    . " WHERE id_fourn = ".$this->idFourn." AND refLdLC = '" . $refLdlc . "'");
+        }
     }
 
     function majPriceFourn($id, $prix, $tva_tx, $ref = null)
@@ -877,6 +883,7 @@ class importCatalogueIngram extends importCatalogueFourn
         'prixBase'        => 8,
     );
     public $sep = ',';
+    public $modeProdFourn = "update";
     
     public function getFileName() {
         return "PRICE.TXT";
